Move collector to guaccollect.
Also add seperate poll option to allow polling file locations, and specific
github repos, depsdev purls, and image metadata. Default to use-csub and poll
true, as this is the service command.
