+++ cmd/guaccollect/cmd/mockcollector/mock_collector.go
