+++ docker-compose.yml
@@ -109,7 +109,7 @@ services:
 
   oci-collector:
     image: "local-organic-guac"
-    command: "/opt/guac/collector image --use-csub"
+    command: "/opt/guac/guaccollect image"
     working_dir: /guac
     restart: on-failure
     depends_on:
@@ -119,7 +119,7 @@ services:
       - ./container_files/guac:/guac
   depsdev-collector:
     image: "local-organic-guac"
-    command: "/opt/guac/collector deps_dev --use-csub"
+    command: "/opt/guac/guaccollect deps_dev"
     working_dir: /guac
     restart: on-failure
     environment:
